refactor(workstation): extract compose + editor actions into hooks (0.72 app.ts decomposition)
